+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Report elements that have cursor pointer style.
+- Report elements with specific `role` attribute values.
+- Include ARIA attributes for uniquely identifiable elements lacking an `id`.
 
 ## 0.1.8 - 2025-12-12